Fix the signed bitwise sites in Tiberian Dawn
551 sites. XYP_COORD is Pixel_Offset_Coord, which casts each component
once and packs unsigned; the cell, coordinate, target and facing helpers
in inline.h and defines.h do the same, with AsFacing wrapping a facing
instead of & 7; the Concrete_Calc overlay indices, the shroud index, the
facing composites, the message segments and the threat mask are unsigned;
layout arithmetic on the resolution factor and on widths is arithmetic;
ring buffers wrap with % size. The house scans go through ScanBit, which
returns 0 for a building type past the 64 bits the scans hold: the last
walls used to wrap onto the first building types' bits.

+++ src/td/mplayer.h
@@ -12,7 +12,8 @@ inline HousesType MPlayerID_To_HousesType(uint16_t id) {
   return static_cast<HousesType>(id & 0x000f);
 }
 inline uint16_t Build_MPlayerID(int c_idx, HousesType htype) {
-  return static_cast<uint16_t>(c_idx << 4 | htype);
+  return static_cast<uint16_t>(static_cast<unsigned>(c_idx) << 4 |
+                               static_cast<unsigned>(htype));
 }
 
 GameType Select_MPlayer_Game();
